+++ migrations/20180112110211_stock_history_ohlc.up.sql
@@ -1,6 +1,6 @@
 ALTER TABLE StockHistory 
 ADD (
 intervalRecord int(11),
-high int(11) ,
-low int(11) ,
-open int(11));
+high bigint(11) ,
+low bigint(11) ,
+open bigint(11));
